Cambios Backend sobre Providers Backoffice:
-  Se ha modificado la BBDD para que providerId sea un INT Auto Incr.
Al no ser un campo AI daba problemas al crear nuevos proveedores.

+++ BackendWT/BackendWT/Controllers/ProvidersController.cs
@@ -14,7 +14,7 @@ namespace BackendWT.Controllers
     {
         public List<Provider> Get() => new ProvidersRepository().GetProviders();
         public bool Post([FromBody] Provider provider) => new ProvidersRepository().SaveProvider(provider);
-        public bool Put(byte providerId, Provider provider) => new ProvidersRepository().ChangeProvider(providerId, provider);
-        public bool Delete(byte provierId) => new ProvidersRepository().RemoveProvider(provierId);
+        public bool Put(int providerId, Provider provider) => new ProvidersRepository().ChangeProvider(providerId, provider);
+        public bool Delete(int provierId) => new ProvidersRepository().RemoveProvider(provierId);
     }
 }
