+++ BackendWT/BackendWT/Models/UserSubscriptions.cs
@@ -8,7 +8,7 @@ namespace BackendWT.Models
 {
     public class UserSubscriptions
     {
-        public UserSubscriptions(int userSubscriptionsId, string providerName, DateTime paymentDate, string billingPeriod, double price, string userId, byte providerId)
+        public UserSubscriptions(int userSubscriptionsId, string providerName, DateTime paymentDate, string billingPeriod, double price, string userId, int providerId)
         {
             UserSubscriptionsId = userSubscriptionsId;
             ProviderName = providerName;
@@ -31,14 +31,14 @@ namespace BackendWT.Models
         [Required]
         public string UserId { get; set; }
         [Required]
-        public byte ProviderId { get; set; }
+        public int ProviderId { get; set; }
         public User User { get; set; }
         public Provider Provider { get; set; }
     }
 
     public class UserSubscriptionsDTO : UserSubscriptions
     {
-        public UserSubscriptionsDTO(int userSubscriptionsId, string providerName, DateTime paymentDate, string billingPeriod, double price, string userId, byte providerId, byte[] providerLogo) : 
+        public UserSubscriptionsDTO(int userSubscriptionsId, string providerName, DateTime paymentDate, string billingPeriod, double price, string userId, int providerId, byte[] providerLogo) : 
             base (userSubscriptionsId, providerName, paymentDate, billingPeriod, price, userId, providerId)
         {
             ProviderLogo = providerLogo;
